+++ functions_framework/lib/src/log_severity.dart
@@ -12,12 +12,39 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+abstract class CloudLogger {
+  const CloudLogger();
+
+  void log(Object message, LogSeverity severity);
+
+  void debug(Object message) => log(message, LogSeverity.debug);
+
+  void info(Object message) => log(message, LogSeverity.info);
+
+  void notice(Object message) => log(message, LogSeverity.notice);
+
+  void warning(Object message) => log(message, LogSeverity.warning);
+
+  void error(Object message) => log(message, LogSeverity.error);
+
+  void critical(Object message) => log(message, LogSeverity.critical);
+
+  void alert(Object message) => log(message, LogSeverity.alert);
+
+  void emergency(Object message) => log(message, LogSeverity.emergency);
+}
+
 /// See https://cloud.google.com/logging/docs/reference/v2/rest/v2/LogEntry#logseverity
 class LogSeverity implements Comparable<LogSeverity> {
+  static const defaultSeverity = LogSeverity._(0, 'DEFAULT');
   static const debug = LogSeverity._(100, 'DEBUG');
   static const info = LogSeverity._(200, 'INFO');
+  static const notice = LogSeverity._(300, 'NOTICE');
   static const warning = LogSeverity._(400, 'WARNING');
   static const error = LogSeverity._(500, 'ERROR');
+  static const critical = LogSeverity._(600, 'CRITICAL');
+  static const alert = LogSeverity._(700, 'ALERT');
+  static const emergency = LogSeverity._(800, 'EMERGENCY');
 
   final int value;
   final String name;